+++ lib/src/unsupported.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 
 ///
@@ -9,8 +11,7 @@ class HighCharts extends StatefulWidget {
       required this.size,
       this.loader = const CircularProgressIndicator(),
       this.scripts = const [],
-      Key? key})
-      : super(key: key);
+      super.key});
 
   ///Custom `loader` widget, until script is loaded
   ///
@@ -95,10 +96,14 @@ class HighCharts extends StatefulWidget {
   ///
   final List<String> scripts;
   @override
-  _HighChartsState createState() => _HighChartsState();
+  HighChartsState createState() => HighChartsState();
+
+  String? getPlatformVersion() {
+    return Platform.operatingSystemVersion;
+  }
 }
 
-class _HighChartsState extends State<HighCharts> {
+class HighChartsState extends State<HighCharts> {
   @override
   Widget build(BuildContext context) {
     return const Center(child: Text("HighCharts: UnSupported Platform"));