+++ lib/src/web/high_chart.dart
@@ -15,8 +15,7 @@ class HighCharts extends StatefulWidget {
       required this.size,
       this.loader = const CircularProgressIndicator(),
       this.scripts = const [],
-      Key? key})
-      : super(key: key);
+      super.key});
 
   ///Custom `loader` widget, until script is loaded
   ///
@@ -101,12 +100,12 @@ class HighCharts extends StatefulWidget {
   ///
   final List<String> scripts;
   @override
-  _HighChartsState createState() => _HighChartsState();
+  HighChartsState createState() => HighChartsState();
 }
 
-class _HighChartsState extends State<HighCharts> {
+class HighChartsState extends State<HighCharts> {
   final String _highChartsId =
-      "HighChartsId" + (Random().nextInt(900000) + 100000).toString();
+      "HighChartsId${Random().nextInt(900000) + 100000}";
 
   @override
   void didUpdateWidget(covariant HighCharts oldWidget) {